+++ src/modules/multi_ir_bgck_mod.f90
@@ -28,7 +28,7 @@ module multi_ir_bgck_mod
   use mpi_mod
   use columnData_mod
   use mpivar_mod
-  use EarthConstants_mod
+  use earthConstants_mod
   use MathPhysConstants_mod
   use verticalCoord_mod
   use presProfileOperators_mod