+++ src/programs/src_files/src_files_ensPostProcess.sh
@@ -1,7 +1,7 @@
 SRC_FILES=""
 SRC_FILES="$SRC_FILES clib_interfaces_mod.ftn90 randomnumber_mod.f90"
 SRC_FILES="$SRC_FILES utilities_mod.f90"
-SRC_FILES="$SRC_FILES bufr_mod.f90 earthconstants_mod.f90 mathphysconstants_mod.f90 mpi_mod.f90"
+SRC_FILES="$SRC_FILES bufr_mod.f90 mathphysconstants_mod.f90 earthconstants_mod.f90 mpi_mod.f90"
 SRC_FILES="$SRC_FILES mpivar_mod.f90 physicsfunctions_mod.f90 varnamelist_mod.f90"
 SRC_FILES="$SRC_FILES horizontalcoord_mod.f90 kdtree2_mod.f90 verticalcoord_mod.f90"
 SRC_FILES="$SRC_FILES codeprecision_mod.ftn90 oceanMask_mod.f90 ramdisk_mod.f90 timecoord_mod.f90"